Optimization

+++ shaders/libs/encoding.glsl
@@ -29,20 +29,20 @@ const float entityFlag = 0.2f;
 const float airFlag = 0.0f;
 const float skyObjectFlag = 0.02f;
 
-bool maskFlag(float16_t f0, float16_t f1) {
+bool maskFlag(float f0, float f1) {
 	return abs(f0 - f1) < 0.005;
 }
 
-f16vec3 normalDecode(f16vec2 enc) {
-	f16vec4 nn = f16vec4(2.0 * enc - 1.0, 1.0, -1.0);
+vec3 normalDecode(vec2 enc) {
+	vec4 nn = f16vec4(2.0 * enc - 1.0, 1.0, -1.0);
 	float16_t l = dot(nn.xyz,-nn.xyw);
 	nn.z = l;
 	nn.xy *= sqrt(l);
 	return normalize(nn.xyz * 2.0 + vec3(0.0, 0.0, -1.0));
 }
 
-f16vec2 normalEncode(f16vec3 n) {
-	f16vec2 enc = normalize(n.xy) * (sqrt(-n.z*0.5+0.5));
+vec2 normalEncode(vec3 n) {
+	vec2 enc = normalize(n.xy) * (sqrt(-n.z*0.5+0.5));
 	enc = enc*0.5+0.5;
 	return enc;
 }
